pyproject.toml
[project]
name = "mamba-kan-anomaly-detection"
version = "0.1.0"
description = "Hybrid Mamba-KAN architecture for interpretable sensor anomaly detection"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "Suhas Jagtap" },
]
keywords = ["mamba", "kan", "anomaly-detection", "time-series", "sensor", "state-space-model"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "torch>=2.2.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "scikit-learn>=1.4.0",
    "scipy>=1.12.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "mlflow>=2.10.0",
    "streamlit>=1.30.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "plotly>=5.18.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "tqdm>=4.66.0",
    "pyyaml>=6.0.0",
    "requests>=2.31.0",
    "loguru>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]
gpu = [
    "mamba-ssm>=2.0.0",
    "causal-conv1d>=1.2.0",
]

[project.scripts]
train = "src.training.trainer:main"
serve = "src.serving.api:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "ANN", "S", "B"]
ignore = ["ANN101", "ANN102", "S101"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
